Use a ring buffer for the sender and up baud to 50
The output data seemed to get truncated when starting and stopping the
stream for each message, so instead, keep the stream running
continuously and feed it data through a ring buffer. This complicates
the logic, but its necessary for increasing the symbol rate.

Also use automatic dependencies in the makefile.