Make sender callback agnostic to message format
The sender callback does too much processing, and adding integrity
checks or forward error correction is way too much. So, instead, divide
the message into symbols in the I/O loop and put that into the ring
buffer. This is less memory-efficient and less convenient for
interactive usage (you have to manually input EOF with Ctrl-D to send a
short message), but it paves the way for expanding the protocol to
support FEC.